feat(section-metadata): migrate to importer pattern, remove nx2 copy
